Rename bmi* variables, and change them to be int32_t
These are integer counts, so int32_t is the more natural type for them.
While the world's population doesn't fit into 32bits, a pixel's worth of
the population surely will.

The increment/decrement operations should now be faster too.

Also renamed the variables, so it's clear what each of them is.

This also means we correctly handle the case where there are more than
16777216 people in a pixel, should that ever come up.