Update MachineDefines.h
This is a port of PR#99 from the private repo to the public one:

Original commit message:

x86 cpus are all 64 byte cache lines, and we might want to run on one of
the big quad socket cloud instances with 96 cores.  This is performance
neutral and a slight memory savings.