Guard uses of mcell_adunits[i] as array indices
mcell_adunits[i] are all initialised to -1.  Later on the values are
updated, and then used to index into other arrays.  However, there is
no check that the index has been set to something other than -1, and so
there may be out of bounds accesses.  This shouldn't happen based on
other conditions - but is not always statically determinable.

This commit adds some checks to ensure that we don't do an out of bounds
access.