Ensure we don't access AdUnits[-1]
The previous commit highlighted that mcell_adunits[i] was still -1 in
some circumstances.  This commit fixes that.

The fix is to condition the access to AdUnits on Mcells[i].n > 0, at
which point mcell_adunits[i] should be >= 0.

This does not change behaviour as I contend that:

mcell_adunits[i] == -1 iff Mcells[i].n == 0

Both indicate that there is no population in an mcell.

This is also why this didn't produce any issues previously because when
mcell_adunits[i] was -1 Mcells[i].n was 0 and so the out of bounds write
did not change the value.