docker: Add Dockerfile and Actions workflow for publishing Docker images.
This adds a Dockerfile to the project to enable easier and more uniform
testing. It also adds an Actions worflow to publish the images for
`master` and tags to the GitHub Package Regesitry so that they can be
more easily consumed by parties that may wish to run the model without
compiling it themselves.