Open all files in binary mode
We want to be able to use the same input files on both Windows and
Linux, so we need to be line-ending-agnostic anyway. So we may as well
be in binary mode, so that we can be sure that if an input file works on
one platform, then it works on all platforms.