Comment CellLookup, and use its result as a pointer directly
We already used it as a pointer a lot, but in some cases we converted it
back to an array index.

There's still a couple of places where we use it to calculate an index,
as we then do some arithmetic with the index to get microcell indices
from it.