Refactor setting of RNG seeds
There were a couple of places where we use
    P.newseed1 = (int)(ranf() * 1e8);
to make new seeds, but Rand's setall already implements RNG splitting,
so we now use that instead.

This needed a few knock-on changes in the places where we seed the RNG,
as if we wanted to reuse the seed values then we now need to store them.

Finally, we now also split the seeds used for the setup phase, and
reinitialise the RNG with them after the network has been created or
loaded. It looks like historically the second pair of seeds had been
used at this point, to make the runs identical regardless of how the
network was made, but that this had been changed when seed-resetting was
implemented. Now the determinism is back, fixing #116.