Fix `nim` computation
It looks like it is intended to count the number of times DoImmune is
called, but (a) in one case it wasn't incremented, and (b) in another
case it was incremented even if it wasn't called, due to not having
braces around the body of an `if`.

I've also uncommented the statement that prints the value out, but put
it in `if (false)`. This will be optimised out by the compiler, but it
let's it see the code, so the variable doesn't look like it is unused.