Use an unbiased shuffle to shuffle PeopleArray
The previous code looks like it suffers from the bias described by
  https://en.wikipedia.org/wiki/Fisher%E2%80%93Yates_shuffle#Na%C3%AFve_method
so we now use Fisher–Yates instead
  https://en.wikipedia.org/wiki/Fisher–Yates_shuffle

While the previous code did appear to make some attempt to make sure
that every element moved, (a) it was possible for an element to move and
then be moved back again, and (b) I can't see anything that elsewhere
that depends on every element having moved. I've therefore assumed that
this isn't required.