When reading a text pop density file, store in memory in the binary format
This means that a lot of the later code can be non-conditional on the file
format, simplifying it.

This also saves us from reading the file twice when using admin units
(once in SetupModel, then again in SetupPopulation). We do still need to
read the contents twice, in order to get the number of lines, but the
first pass doesn't need to do the slow sscanf parsing.