Set -fp-model precise when using the ICPC+Linux
Set the floating point model to "precise" rather than "fast" which is the default when compiling with the Intel compiler on Linux.  This allows the code to pass the regresson tests on this platfom (assuming the value of `OMP_NUM_THREADS` is high enough for the multi-threading test to pass!).